Break out Command and Externs into modules
adds a few fields to the cabal file
