+++ psc-ide.cabal
@@ -15,6 +15,8 @@ cabal-version:       >=1.10
 library
   hs-source-dirs:        lib
   exposed-modules:       Purescript.Ide
+                       , Purescript.Ide.Command
+                       , Purescript.Ide.Externs
   default-language:      Haskell2010
   build-depends:         base >= 4.7 && < 5
                        , text
@@ -49,10 +51,12 @@ executable psc-ide-server
                      , filepath
 
 test-suite spec
-  type: exitcode-stdio-1.0
-  ghc-options: -Wall
-  hs-source-dirs: test
-  main-is: Spec.hs
-  build-depends: base    >= 4.7 && < 5
-               , psc-ide
-               , hspec
+  type:                exitcode-stdio-1.0
+  ghc-options:         -Wall
+  hs-source-dirs:      test
+  default-language:    Haskell2010
+  main-is:             Spec.hs
+  other-modules:       Purescript.Ide.CommandSpec
+  build-depends:       base    >= 4.7 && < 5
+                     , psc-ide
+                     , hspec
